test the existence of ~/.TinyTeX/bin (because Travis's cache will create ~/.TinyTeX)

+++ tools/download-travis-linux.sh
@@ -4,9 +4,11 @@ mkdir -p ~/bin
 
 [ ! -z $TINYTEX_FORCE_INSTALL ] && rm -rf ~/.TinyTeX
 
-  wget -q https://travis-bin.yihui.name/tinytex.tar.gz;
-  tar xzf tinytex.tar.gz -C ~/;
-  rm tinytex.tar.gz;
+if [ ! -d ~/.TinyTeX/bin ]; then
+  wget https://travis-bin.yihui.name/tinytex.tar.gz
+  tar xzf tinytex.tar.gz -C ~/
+  rm tinytex.tar.gz
+fi
 
 cd ~/.TinyTeX/bin/*/
 ./tlmgr path add
